Reorganizing options into their own extension
