+++ Sources/OAuthKit/Extensions/Date+Extensions.swift
@@ -17,5 +17,4 @@ public extension Date {
     static func - (lhs: Date, rhs: Date) -> TimeInterval {
         return lhs.timeIntervalSinceReferenceDate - rhs.timeIntervalSinceReferenceDate
     }
-
 }
