+++ Sources/OAuthKit/OAuth+Request.swift
@@ -258,7 +258,6 @@ extension OAuth {
             return queryItems
         }
     }
-
 }
 
 fileprivate extension URLQueryItem {
