+++ Sources/OAuthKit/OAuth+Token.swift
@@ -60,5 +60,4 @@ extension OAuth {
             case openIDToken = "id_token"
         }
     }
-
 }
